Create a WorkerStartStopLogger to encapsulate worker events logging, moving worker_start and worker_stop from from AdminLogger.
